# src/main/java/com/example/demo/core/domain/MetricItemResponse.java
package com.example.demo.core.domain;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * Model
 */
@Data
@AllArgsConstructor
@NoArgsConstructor
public class MetricItemResponse {
    private String request;
    private Integer count;
}