# src/main/java/com/example/demo/core/domain/MetricListResponse.java
package com.example.demo.core.domain;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * Model
 */
@Data
@AllArgsConstructor
@NoArgsConstructor
public class MetricListResponse {
    private MetricItemResponse[] metrics;
}